ui: add javascript, css, and template for main page
The idea behind this page is that it acts as three IconViews that
represent the VMs, any defined templates, and a miscellanous view
for creating new templates, etc.

There is a host menu for selecting host options, a tab for federation,
and an action menu for each IconView.

The Icons in each IconView are meant to be live tiles of the running
VMs.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>