Add basic model/controller support
This patch introduces the model/controller architecture to burnet:

The model implements the "backend" operations.  It is the source of information
about virtual machines and other API concepts.  The model does the real work
(such as creating a new VM in the system).  Right now we only have a Mock model
which provides dummy data and where operations are mostly no-ops.  It is enough
to be a good test bed for the UI though.

Controllers make the backend functionality available via a RESTful API.
Resources and Controllers are assembled into a tree-like structure such that
their relationships are represented using URIs.  For example: The Root Resource
forms the root of the tree and its URI is '/'.  Eventually we will add a
VirtualMachines Collection to the Root Resource and its URI will be '/vms'.
Controllers do no real work on their own but instead rely on the model for this.

Since Collections are very similar to one another and Resources all share a
common set of base functionality, I have established two parent classes
'Collection' and 'Resource' which can be extended when we expand the API.  This
avoids boilerplate code and prevents mistakes (such as not raising HTTP:404 if a
Resource cannot be found).

Signed-off-by: Adam Litke <agl@us.ibm.com>
Message-id: 1364247685-31132-1-git-send-email-agl@us.ibm.com
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>