Implement the real backend using libvirt
This patch adds our "real" backend which connects to libvirtd to
manipulate real Virtual Machines.  VMs are persisted by libvirt but
since our concept of VM Templates is not part of the libvirt API we will
need to persist these within Burnet.  The next patch will implement a
simple form of object persistence but for now we will just use a
dictionary that loses state between server invocations.

Reviewed-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>