Add a simple object persistence mechanism
V3 -> V4:
 - Tests: have more tests use a temporary object store

V2 -> V3:
 - To avoid race conditions, limit access to one thread at a time
 - Convert ObjectStore into a python ContextManager to handle locking
 - Addressed misc review comments (sming)

The Libvirt API is able to manage persistence for its objects (such as
VMs and Storage Pools).  For Burnet objects that require persistence
(eg. VM Templates), we need to manage persistence ourselves.
Fortunately, a simple mechanism should be good enough.

We need a way to store and retrieve simple objects with the following
methods:
 - get_list(type): Get a list of items of the given type
 - get(type, name): Get an object of the given type and name
 - store(type, name, obj): Store a an object of the given type and name
 - delete(type, name): Remove an object of the given type and name

Objects are stored as json strings in a sqlite3 database.  I don't
anticipate the need for complex database operations, but should that be
necessary in the future, we can expand this.

Access to the ObjectStore is given in the form of an ObjectStoreSession.
When the ObjectStore class is used as a context manager (the with
statement), a session is provided to the caller.  within that session,
the caller has locked, exclusive access to the ObjectStore.

Threading note: The sqlite3 module does not support sharing connection
objects between threads.  Therefore, the ObjectStore class maintains one
connection per thread.  We only keep around 10 connections at any given
time.  This will allow us to reclaim the memory from connections created
by old cherrypy threads.

Reviewed-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>