ui: Allow scrollbars on the VNC window if necessary
If the VM's screen resolution is higher than the resolution of the
client, the VNC viewport will only be partially displayed.  This makes
it impossible to work with the VM properly.  Add scrollbars=yes to the
vnc popup javascript call to allow scrollbars to be added to the window.

On Firefox, the scrollbars only appear if the window is too large to
completely fit on screen.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Message-id: 1368794794-31094-1-git-send-email-agl@linux.vnet.ibm.com