ui: Filter all requests for html pages through cheetah
Cheetah is the mechanism we use to perform translations.  All
translatable text is declared in html pages.  Translations in javascript
are currently not allowed or supported.

This patch alters the Root Resource so that all requests for html files
will be routed through cheetah for translation.

Dynamic data (such as hostnames) should not be passed into cheetah
anymore.  All dynamic information must be retrieved by the client via
the REST API.  For the dev ui, this means we must stop passing in
hostname.  For now I have hard-coded it to 'localhost' but once we have
the federation APIs, this host's name can be retrieved via the API.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Reviewed-by: Shaohe Feng <shaohef@linux.vnet.ibm.com>