ui: Namespace the dev ui
We want to keep the dev UI in the tree for now but to avoid naming
conflicts with the design UI when it is merged, prefix all dev ui files
with 'dev.'  Also, flatten the css, images, js, etc. directories.  We'll
be handling theming by adding themes to the design ui.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Reviewed-by: Anthony Liguori <aliguori@us.ibm.com>