Update the Rest and MockModel tests to account for the server no longer redirecting
tests updated to no longer expect 303 response. The screenshot test update is the most
interesing. It previously got the redirect URL and confirmed that 5 seconds later it
got the same redirect url. Now it fetches the screenshot url to ensire there is a
screenshot avail, then gets the VM info which currently includes a direct link to the
screenshot. After 5 seconds it refetches the screenshot url and confirms that the image
is still the same.

Signed-off-by: Adam King <rak@linux.vnet.ibm.com>