fix bug: let i18n_files depends on po files instead of mo files.
Git clone the burent source, and make burnet package first time, the i18n
mo files are missing. So they can not be copied to the build package during
building time.

Adam suggested to create a new variable i18n_languages with the current set
of languages and use that to construct the file list.

accord with Adam's suggestion, this patch does avoid a list comprehension
that is very difficult to comprehend.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-by: Royce Lv<lvroyce@linux.vnet.ibm.com>