fix bug: UI displays abnormal follow the README
For the python setup.py can not find theme-default.min.css and
burnet.min.js

Adam suggestion:
we should just explicitly list the files we want to include.
It's only a few files and the only time it would change is if
a new theme is added.

Agree with Adam, a explicitly list is the simple way to avoid this problem

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-by: Royce Lv<lvroyce@linux.vnet.ibm.com>