doc: Update README.md with more package dependencies
libvirt and libvirt-python are also required in order to run kimchi.
Update the README with this information.

libvirt is not necessarily.  Libvirt has a remote protocol so it is possible
to just install the libvirt library and client utils (python bindings) but
not install the local daemon.

Here we want to explicitly select the local libvirt daemon.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>