Issue #59: Make the popup menu entirely visible in browser
In the guest list page, when some VM line is at the bottom of the
browser, and the user clicks "actions" button, the popup menu may
go out of the viewport or just show a part. We need to make sure
the whole body of the menu is in the viewport.
    In this patch, we did a check to the popup menu and scroll it
into browser viewport.
    Worked in IE9+, Chrome, Firefox 17.0.1, and Safari.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>