Issue #43: UI Reports Error When Pressing Enter Key in Creating VM Form
Browser redirects to an error page when pressing Enter Key in Creating
VM Form. It's the default behavior of an HTML form when pressing Enter
Key to submit the form. However, we do not want to submit it directly.
Instead, we use Ajax to submit it to make our UI consistent.

    In this patch, we disable the default behavior of an HTML form. To
always return false; will prevent the form from being submitted in the
submit event listener.

    Tested on IE9, IE10, Firefox 22, Chrome, and Safari.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>