template-update: mock model support implement update method
implement template_update by calling template_delete followed by
template_create.
As Adam Litke suggested: save the original template so update will
rollback if something fails.

In mockmodel the info of template is stored in memory.
So the update can be this process: lookup -> copy -> delete -> create.
But this process can not by apply to VM create.
There is some specific data for a VM and we can't simply delete it and
then recreate it.

We should support this template update first, and then support VM update.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>