set expires for static content caching
As Adam King suggested:
We don't think cache expires is a value that we will want a typical
admin changing as they would the other values in kimchi.conf.
so put the explicit value to just one place in the python code,
and comment that its a year, and have the other code reference it.
If we ever want to change it, we just need to update that one value.

after set the expires, the cilent can receive a response-head as follow.
response-head:
	HTTP/1.1 200 OK
	Content-Length: 55946
	Accept-Ranges: bytes
	Expires: Sun, 17 Aug 2014 08:27:06 GMT
	Server: CherryPy/3.2.2
	Last-Modified: Sat, 17 Aug 2013 08:26:23 GMT

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Signed-off-by: Xin Ding <xinding@linux.vnet.ibm.com>
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>