A module to check the pam authenticate
This authenticate base on PyPAM.
It is used to verify userid/pwd on the host system.
PyPAM is a universal package, it can be available on most distribution.
It can be available on el4~el6, f14~f19
http://pkgs.fedoraproject.org/cgit/PyPAM.git/

It is the same interface with python-pam.
but python-pam is not be available on rhel by pam
usage:
    from auth import authenticate
    authenticate(username, passwd)

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Signed-off-by: Shu Ming <shuming@linux.vnet.ibm.com>