Catch the libvirt exceptions when accessing storage objects
Raise the kimchi defined exceptions and report the exception messages to controller,
then user can get useful information of the libvirt exceptions instead of '500' HTTP error.

Use 'get_class_name()' in Resource index() method to define the message is not accurate
when accessing resource of sub-collection if sub-collection not found, so we just forward the original
message.

Signed-off-by: Bing Bu Cao <mars@linux.vnet.ibm.com>