Enable REST API
This patch depends on:
 - [PATCH 0/3 v4] SSL Support
 - [PATCH V2] A module to check the pam authenticate

With SSL and PAM authentication support present, we are now ready to
protect our website with a login screen.  Here's how it works:

We enable a new tool 'kimchiauth' on most of the content served by
kimchi.  This includes the REST API and UI pages.  The login page is the
only thing that requires no authentication.  For now, some JS and CSS
are also allowed without a login (this allows the login page to access
some needed resources).

The kimchiauth tool checks if there is an active session before serving
the content.  If there is, then the content is served as normal.  If
there is not then the following happens:  If the REST API is being
used, HTTP Basic Auth headers are checked.  If present, the provided
credentials are used to attempt a login.  If this works, the content is
sent as normal.  If the UI is being used, the login page is displayed.
When the user enters their credentials on the login page, a call is made
to /login with the given credentials.  On the server, /login will
attempt to authenticate with the supplied credentials.  If successful, a
session is established and the client will reload the original URL.

This scheme has some notable properties:
 - HTTP Basic Auth can be used by REST API consumers.  It's easier than
   trying to use /login.
 - The UI will not use Basic Auth.  This allows for a secure logout
   since the browser will not be caching any credentials.
 - The established session can be used to store user preferences (such
   as preferred language).

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>