auth: Prevent HTTBA in browser with FF Poster
When using Firefox to connnect to the kimchi UI and using the Poster
extension to issue REST requests the browser will start sending the auth
header from both Poster and the main browser.  This causes Firefox to
re-establish the login immediately after Logout is clicked.

This patch uses the X-Requested-With browser detection idea to screen
out HTTPBA authentication if the browser is requesting it.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>