Issue #125: Switching Page Tabs Breaks Pop-up Menu Disappearing
V2 -> V3:
  3a) Rebase.
  3b) Add commit message.

v1 -> V2:
  2a) Rebase.

Pop-up menu depends on document's mouse click event to disappear.
In former implementation, click event listener was bound and will
be triggered. The problem is that "return false" will prevent the
click event to bubble to document node. In this patch, we'll take
"event.preventDefault()" to avoid <a href> redirecting browser to
other pages and still bubble click event to document.

    Tested on IE9, IE10, Firefox 22, Chrome, and Safari.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>
Reviewed-by: Meina Zhou <zhoumein@linux.vnet.ibm.com>