Create new thread to collect Kimchi statistics
That way Kimchi statistics is collected independent of server resquests.

Use cherrypy.process.plugins.BackgroundTask to create the new thread to avoid
problems while stopping the server.

Also reorganize how the data is exposed in order to avoid a lot of stats
variables as more statistics will be added to kimchi.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>