Issue #151: No Message when User Opens http://kimchi:port/#i18n
CUrrently we'll concatenate the hash part with ".html" to load the
page into main area. It makes no sense for some HTML pages like:
  * guest.html
  * guest-add.html
  * i18n.html
  * template-add.html
  * template-edit.html
  * more in the future...

  So we need to tell the user it's an invalid URL. We'll check the
page URL firstly before load the page that: if the page is not one
of the href in navigation tab, we'll not load it; instead, we will
redirect to the home page.

  As a consequence, page that does not exist will also be captured
by this code and 404 listener for HTML Ajax request  will never be
triggered. We remove this code, too.

  Tested on Chrome, Firefox, Safari, IE9, and IE10.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>