Fix bug Regex in checking url
v1-v2
1 Make each regex a var to make it more clear
2 check the url by ip type and domain type
There is an issue in the Regex of checkout url
of missing "+/?" in the end of regex.That make some url very long
can not passed by checking. So send a patch to amend this.To make
url just like:http://download.fedoraproject.org/pub/fedora/linux/
releases/19/Live/x86_64/Fedora-Live-Desktop-x86_64-19-1.iso can be
passed in the front end.
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>