Issue #174: error page can not be translated, fix it.
we do not list error.html.tmpl in po/POTFILES.in for a known issue.
issue: https://github.com/kimchi-project/kimchi/issues/174

Wsgiserver will encode the error.html to 'ISO-8859-1'.
But the error.html from cheetah template is unicode.
So we should encode unicode to string before the Wsgiserver
handle it.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>