update test case after a series unit bug fixed
for test_vm_lifecycle and test_vm_on_alt_storage test case:
The size unit of volume is GB when it is created by template create API.
And we get the 'capacity' unit from volume get API is Bytes.
So we should verify it like:
        self.assertEquals(1 << 30, vol['capacity'])

for test_iso_scan_shallow test case:
And the size unit of volume is MB when the volume is created by volume
create API.

It is not a good test case to the 'allocation' and 'capacity' are equals
from from volume get API. It is makes no sense.
self.assertEquals(storagevolume['allocation'], storagevolume['capacity'])
bug: https://github.com/kimchi-project/kimchi/issues/177

we should verify 'allocation' and 'capacity' separately that they are expected.

for test_storagevolume test case:
we should be care the 'capacity' units are different.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>