Drop the handle in storage,fix the position error of the navbar arrow
1 Drop the handle in storage
the blue handle is useless in the storage page, and will mislead the
user to click that handle bar. So drop that blue handle to make the
page more clear.

2 Change the init position of the navbar arrow.
The position of navbar arrow is the right left of the page, but not
the guest page. So change it to guest page for default.

Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>