Issue #190: Language Selector Value NOT Match Page Content
When there is no language cookie set, and the browser language is set
in non-English, language selector in log-in window may have incorrect
value with regard to the page's content. Here is an example:
1) User sets browser language to Chinese (zh_CN);
2) User types Kimchi URL;
3) Server notices request header "Accept-Language" is set to "zh_CN",
   and returns page content with Chinese;
4) JS notices there is no language setting in cookie so no options in
   language selector will be selected and language selector will show
   its first option "English (en_US)".

  The problem lies in Step 4). In this patch, we correct it by adding
a check to the "lang" property of <html> DOM node. With the value, we
can keep language selector's value be consistent with page content.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>