Make header text sizes equal
The list title CSS classes on the pages "Guest" and "Storage" defined
different font sizes for the header texts. Those texts are similar, so they will
look better if they have the same size. This patch applies the CSS class
"list-title" to the text on the page "Storage", so it looks the same as
the text on "Guests", and removes the CSS class "storagelist-title" (and
its subclass) because it did not define anything different than
"list-title".

Fix issue #192 (Header texts have different sizes).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>