Issue #199: Update templates list only after success of deleteTemplate
After deleting a template, the page "Templates" may reload itself to display the
updated templates list even if the delete operation hasn't yet been finished.
In that case, the list update procedure starts before the template is actually
removed from Kimchi's database, and that template is not removed from the view.

Instead of updating the templates list immediately after triggering the delete
operation (without waiting for it to finish), update it only after the delete
operation has finished successfully.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>