feature tests: Verify QEMU can properly work with hostname while streaming ISOs
The ISO streaming functionality is not working well on Fedora 19.
With qemu-kvm-1.4.2-12 and curl-7.29.0-6 packages installed, it is
not able to create virtual machines with streamed ISOs when the
cdrom URL uses host name instead of IP address.
So create a sanity test in order to identify that case to properly
handle it on Kimchi.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>