Add json schema for API's request parameters
This patch adds a json schema defintion for vm's create operation
and validates the request parameters agaist the schema when creating
a new VM by calling jsonschema's validator. So it's safe to remove
the parameter validation in model level.

It also adds two test cases for it. And more schema definitions will
be added in the following patches.

For json schema specification, please see:
	http://json-schema.org/documentation.html

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>