Expose if the host has at least one system report tool
This is needed to enable/disable system reports on UI.
When the host does not have at least one system report tool the user
will not be able to generate debug reports.
For it, the current debug reports implementation was splitted into small
functions to avoid duplicating code but the functionality remains as it was.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>