network: update controller
define Networks collection and Network resource according to API.md

update API.json

Do not add a pattern for subnet parameter.
we can define it as follow:
192.168.0.1/24 or 192.168.0.1/255.255.255.0
"^((2[0-4]\d|25[0-5]|[01]?\d\d?)\.){3}(2[0-4]\d|25[0-5]|[01]?\d\d?)/
(((254|252|248|240|224|192|128|0)\.0\.0\.0)|
 (255\.(254|252|248|240|224|192|128|0)\.0\.0)|
 (255\.255\.(254|252|248|240|224|192|128|0)\.0)|
 (255\.255\.255\.(254|252|248|240|224|192|128|0))|
 (30|[1-2]\d|[2-9]))$"

A invalid parameter {"subnet": "192.168.0.1/255.255.255.1"},
json-schema will report an error like this:
Invalid parameter: '192.168.1.123/34' does not match
'^((2[0-4]\\d|25[0-5]|[01]?\\d\\d?)\\.){3}(2[0-4]\\d|25[0-5]|[01]?\\d\\d?)/
(((254|252|248|240|224|192|128|0)\\.0\\.0\\.0)|
 (255\\.(254|252|248|240|224|192|128|0)\\.0\\.0)|
 (255\\.255\\.(254|252|248|240|224|192|128|0)\\.0)|
 (255\\.255\\.255\\.(254|252|248|240|224|192|128|0))|
 (30|[0-2]?\\d))$'

This message is not friendly to user.

We can use ip-address format in draft3 for IP, that is same with ipV4 in
draft4.
But it does not work.
The json-schema doc says:
format support is optional, and as such other tools may ignore this keyword:
this can lead to a different validation outcome for the same data;

I also try ip-address format, it do can not work.

Also there is not a format to verify netmask in json-schema draft3 and draft4.
So json-schema can not easy to verify subnet.

json-schema doc says:
Not all constraints can be expressed.

So these checks are left to kimchi applications.
And python netaddr can verify subnet easily.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>