Packaging: add the disks.py to src/kimchi/Makefile.am
src/kimchi/disks.py is for listing host partitions and is imported in
src/kimchi/model.py. disks.py is not in src/kimchi/Makefile.am, so
make install, make dist and make rpm do not install or package this
file. You get an import error when kimchi is run from the installed
directory tree.

This patch adds disks.py to Makefile.am and sovle the problem.

v2: Adjust the listing according to alphabetic order.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>