bug fix: Don't use shell=True in subprocess.Popen while running qemu-io test
On Ubuntu 13.10, when using subprocess.Popen with shell=True two processes will
be created:

1) /bin/sh -c qemu-io http://localhost:8000/images/icon-fedora.png -c 'read -v 0 512'
2) qemu-io http://localhost:8000/images/icon-fedora.png -c read -v 0 512

Special in this case, qemu-io command hangs (because it can't resolve hostname)
and we kill the process to interrupt the operation but in this case proc.kill()
will only kill the first process and the second one will keep alive consuming
memory.
Also with the second process alive we can complete feature tests as it waits the
qemu-io command to finish. And then we get the following error while trying to
stop kimchi.

Traceback (most recent call last):
  File "./src/kimchid", line 86, in <module>
    sys.exit(main(sys.argv[1:]))
  File "./src/kimchid", line 83, in main
    kimchi.server.main(options)
  File "/home/alinefm/kimchi/src/kimchi/server.py", line 254, in main
    srv.start()
  File "/home/alinefm/kimchi/src/kimchi/server.py", line 245, in start
    cherrypy.engine.start()
  File "/usr/lib/python2.7/dist-packages/cherrypy/process/wspbus.py", line 235, in start
    self.publish('start')
  File "/usr/lib/python2.7/dist-packages/cherrypy/process/wspbus.py", line 197, in publish
    output.append(listener(*args, **kwargs))
  File "/home/alinefm/kimchi/src/kimchi/model.py", line 224, in _set_capabilities
    self.qemu_stream_dns = FeatureTests.qemu_iso_stream_dns()
  File "/home/alinefm/kimchi/src/kimchi/featuretests.py", line 110, in qemu_iso_stream_dns
    thread.join()
  File "/usr/lib/python2.7/threading.py", line 947, in join
    self.__block.wait()
  File "/usr/lib/python2.7/threading.py", line 339, in wait
    waiter.acquire()
KeyboardInterrupt

So use shell=False to avoid creating two processes and then proc.kill() will not
leave any one running.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>