Issue 277, Disable 'bridged' option when no interface available
Signed-off-by: Yu Xin Huo <huoyuxin@linux.vnet.ibm.com>

The backend behavior has changed when returning interfaces.
If a certain inteface is already configured for a network,
then it will not be returned in interfaces request.
This will run host into a situation that no inteface available.

To get UI more user friendly for this case,
Disable 'bridged' option when no inteface available any more.