storagepool: refactor _get_pool_xml()
In src/kimchi/model.py, _get_pool_xml() is to generate libvirt storage
pool XML definition from arguments provided by the POST data (a JSON
dict). It has to support various types of pool such as dir, netfs and
logical. Now it uses "if ... elif ... elif" to check the requested type
of pool and go in to the respective branch and generates the correct
XML. It also performs some preparations such as create target directory.
This is OK for now but in future we'll have more types of pool, and add
more validation and preparation code. So it needs a little refactor.

In this patch we define StoragePoolDef object along with a factory
methoed named "create". The factory method instantiates DirPoolDef,
NetfsPoolDef or LogicalPoolDef according to the requested pool type.
Then we can call defObj.prepare() to check NFS mount and iSCSI
login/password. We can inspect defObj.xml to get the generated XML for
the respective poolType.

v2 -> v3
  Call pool.build() for netfs, dir and logical pool. So do not need to
  create dirs any longer. Since NFS and iSCSI pool need validation,
  extract a StoragePoolDef object with "prepare" and "get_xml" methods.

v3 -> v4
  Turn 'defObj.get_xml()' into a lazily calculated property 'defObj.xml'.

v4 -> v5
  Do not cache generated xml, have subclass of StoragePoolDef override
  the '.xml' property. This makes the interface cleaner.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>