qemu user tests: probe the username of qemu process started by libvirt
We want to fix the searchable permission for qemu user.
But we should find the username of qemu process firstly.

searchable permission is a known problem.
We have discussed several times on IRC.

Royce reports the qemu username is different on different distros

Adam, Aline, Royce and I think we can probe qemu username with the
similar method of qemu iso stream support.

Zhou Zheng Sheng gives a better way to find the qemu process ID

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Signed-off-by: Royce Lv <lvroyce@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>