template supports networks: let template xml support more networks
If a VM is just in one network, it may not access other machines in
different networks.

In most product environment, more networks are needed.

Such as a network management, a service manages all the VMs in
one network. Such management system is isolated from the internet
for security reasons.

And a more public network is needed for a VM to access the internet.

So a vm with more networks is a common scenario.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>