Github issue #300: hardcoded server port fix
When executing kimchi using a port different than 8000, the "Remote ISO Image" will always be disabled.

The issue is caused by functions in featuretests.py that uses the port 8000 hardcoded, instead of getting the actual used port to launch the server.

As of now I haven't found any other occurrence of the "8000" hardcoded port anywhere in the code.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>