Issue #308: Use correct iSCSI package for openSUSE
Kimchi depends on an iSCSI system package to validate the iSCSI devices
when creating a storage pool. Currently, the package "iscsi-initiator-utils"
is listed as a dependency, but it is not available on openSUSE. Therefore, the
Kimchi RPM file cannot be installed on that distro.

Update the iSCSI dependency package name for openSUSE to "open-iscsi".

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>