Ubuntu: Add the LVM dependency package to README
The README file contains instructions to build Kimchi from the source
tree, including the exact command to install the dependencies on Ubuntu.
But that command does not contain the LVM package which is needed to probe LVM
devices when creating a logical storage pool.

Add the LVM package ("lvm2") to the apt-get install command in the README file.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>