Issue #310: Query correct name to validate logical devices
The dialog to create a new logical storage pool must validate the user
input by checking if they selected at least one of the available
devices. However, the code is looking for checkboxes with a non-existing
element name and it is always reporting that the user did not select any of
them - even when they do. Therefore, it is impossible to create a
logical storage pool.

Change the HTML element name of the devices' checkboxes on the storage pool
page to match the validation algorithm. Instead of "source.devices", use
the name "devices" on those elements.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>