storagepool: Do not assume different 'else' values in pool type
Currently, there are four different storage pool types. The JavaScript
code that handles that logic assumes different values in the 'else'
blocks of different functions. It is a good practice to be explicit on
what we are checking for and not to depend on the possible last value
out of a set. This can easily lead to bugs in the future if additional
storage pool types are added. Also, the code is harder to follow because
the 'else' assumptions on similar comparisons are sometimes one value
and sometimes another value.

Use explicit and consistent values when checking the different
pool types in 'if/else' clauses.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>