Add a method to fix read permission on ISO files
When starting a virtual machine, libvirt changes the user:group of the disk
files to libvirt-qemu or qemu (depends on Linux distribution)
That way libvirt can work properly with those files.

But if the file does not have read permission, libvirt won't be able to
access them.

This patch adds a method to add read permission on files to avoiding
problems when starting a virtual machine.

Also remove useless imports from utils.py

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>
Signed-off-by: Royce Lv <lvroyce@linux.vnet.ibm.com>