CSS: Do not allow lines longer than 79 characters
In order to make the code more consistent, set a maximum of 79
characters per line in CSS files. If the CSS properties line is longer
than 79, break it and align the second line to the beginning of the
properties of the first line. If the CSS filter line is longer than 79,
break it and do not indent the second line (so it also keeps aligned to
the first line as well).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>