HTML: Make the form element's labels clickable
A good usability feature of an HTML form is to click on an element's
label to activate the element itself. For example, when clicking on a
checkbox's label, the checkbox itself is checked, instead of requiring
the user to click on the small corresponding element.

Assign every checkbox's and radio's labels to their corresponding
form elements.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>