refactor exception: Set error messages for jsonschema validation
While validating data with jsonschema it raises English messages.
To allow translate all error messages to the supported languages it is
needed to set a custom error message.
In this case, the message code is being used. Then KimchiException can
get the correct message based on code and display the translated
message.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>