Add disks to LVM pool: mockmodel changes
Eliminated the parameter validation that is now being done
in API.json.

To simulate the operation of add disks in the mockmodel, for
each disk (which is validated with real data from the host),
for each new disk added a fixed amount is added in the
'capacity' and 'available' fields on the pool.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>