Add software update action to Host resource
Instead of handling the update packages action in
PackagesUpdate(Collection) move it to Host(Resource) to avoid
misunderstanding about the URI.

From /packagesupdate/update URI the "update" can be understood as the
resource item instead of the action.
So use /host/swupdate to update the packages marked to be updated.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>