bug fix: Do not probe ISO file while checking template integrity
When a template is created, the ISO file is probed to get information
about it: if it is bootable, OS name, OS version, etc.

While checking template integrity we just need to know the ISO file
is still accessible. Which can be done by os.path.isfile() to check
local ISO files and kimchi.utils.check_url_path() for remote ones.

That way we don't need to read the ISO content to check ISO information which
is causing a high latency issue when there is a lot of remote templates
created in kimchi.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>