Github bug #307: add storage pool type list - JS changes
The changes in ui/js/src/kimchi.storagepool_add_main.js were made
to relieve the storage pool type dropbox from the return of the
ajax call to retrieve the available partitions of the host.

The dropbox will always have all the options displayed. In case
of the LOGICAL pool type the user won't be able to proceed with
the creation of the pool if no partitions are found. A message
is being displayed to warn the user.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>