storage volume: fix xml parsing of logical volume format type
Not all types of libvirt storage volume provide format information.
Particularly, the volume in logical pool doesn't have format information
in the dumped XML description. StorageVolumeModel.lookup assumes there
is a format tag, but for logical pool volumes it does not exist. So an
exception is raised when it lookup a logical volume.

To trigger this problem, firstly create a guest using a logical pool. Then
click the logical pool from the "storage" tab, and it would fail to list
the volumes.

This patch catches the exception when parsing the format tag, and
returns an 'raw' for the format attribute if there is no available format
information in the XML.

v2
Return 'raw' instead of return empty string, because the only supported
volume format of a logical pool is 'raw'.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>