Support ppc64 isos in Templates: Add arch and urls to distro jsons
If Kimchi is running in a PPC machine, user can create a template
that uses a remote x86_64 iso.
Adding an os_arch field to the distro so we can filter distros
on systems with other architectures. This prevents a user from
creating unusable templates.
This patch also changes the name field of Fedora distros.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>
Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>