bug fix: "sudo: sorry, you must have a tty to run sudo".
The authorization code executes some sudo commands in backend to
check if the user used to login in Kimchi has permission to execute
privileged commands.

By default, sudo needs a TTY to be executed and when Kimchi is executed
as a service by systemd (systemctl start kimchid.service), it runs
disconnected from any TTY, and sudo returns an error and the message:
"sudo: sorry, you must have a tty to run sudo".

This patch sets in Kimchi service file for systemd the /dev/tty11 as
default TTY to execute the service.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>