fix: Build new DITA pages when running "make"
When the user types "make" for the first time, all the .dita files are
converted to .html. But if some new DITA page is added to the directory,
it is not build. The user has to remove all HTML files and run
"make" again.

Update the Makefile so it can build new DITA files if any is added

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>