Pass ajaxError information on to original requester on ajaxError event
The previous implementation of the ajaxError event did not pass along all the
available information to the original caller.  Corrected with this patch.
Display a message if the error is requestRejected.
Updated guest vms processing to use the additional information.
Corrected the error handler to show the JSON response reason when available.

Signed-off-by: Adam King <rak@linux.vnet.ibm.com>