Don't allow templates to be created with ISOs that won't be usable.
qemu can't use an ISO if it has no search permission on the directories
containing the file.
Return an error explaning this, and suggest the way to fix it.

http://libvirt.org/drvqemu.html#securitydac
In the "session" instance, the POSIX users/groups model restricts QEMU
virtual machines (and libvirtd in general) to only have access to
resources with the same user/group ID as the client application. There
is no finer level of configuration possible for the "session" instances.

If QEMU virtual machines from the "system" instance are being run as
non-root, there will be greater restrictions on what host resources the
QEMU process will be able to access. The libvirtd daemon will attempt to
manage permissions on resources to minimise the likelihood of
unintentional security denials

Any files/devices used as guest disk images must be accessible to
the user/group ID that QEMU guests are configured to run as. The
libvirtd daemon will automatically set the ownership of the file/device
path to the correct user/group ID. Applications / administrators must be
aware though that the parent directory permissions may still deny
access. The directories containing disk images must either have their
ownership set to match the user/group configured for QEMU, or their UNIX
file permissions must have the 'execute/search' bit enabled for
'others'.

The simplest option is the latter one, of just enabling the
'execute/search' bit. For any directory to be used for storing disk
images, this can be achieved by running the following command on the
directory itself, and any parent directories

    chmod o+x /path/to/directory

    In particular note that if using the "system" instance and
attempting to store disk images in a user home directory, the default
permissions on $HOME are typically too restrictive to allow access.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>
Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>