bug fix: Identify update tool based on available system tools
Instead of using platform.linux_distribution() and keep a list of possible
values to match, this patch verifies the tools on the system to identify
which update tool to use in Kimchi.
So try to import yum module, apt module or run the zypper command and in
last case raise an error.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>