bug fix: Identify repository management tool based on available system tools
Instead of using platform.linux_distribution() and keep a list of possible
values to match, this patch verifies the tools in the system to identify
which repository management tool to use in Kimchi.
So try to import yum module or apt_pkg module and in last case raise an error.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>