Do not allow user disable/delete a network used by VM or template
Add a new parameter to network resource to identify the network is in
use by a template or VM, and do not allow user disable/delete it in this
case.
The 'default' network is always 'in use' as Kimchi uses it when creating a new
template.

Also update API and test cases to reflect this change

Signed-off-by: Aline Manera <alinefm@br.ibm.com>