Do not show success messages
When the user performs some operations (like updating guest CD-ROMs), a green
success message is displayed on top of the screen. This behavior is not common
in Kimchi; when an operation performs successfully, nothing is shown.

In order to make Kimchi more consistent, remove the code that displays
success messages.

This patch enforces what was applied by 57f9f9f.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>