Skip 'screendump' for vms no video device configured
It's observed that if a vm is created by external tool and doesn't
have video device, libivrt will keep reporting the following error
message:

	"internal error: unable to execute QEMU command 'screendump':
	 There is no QemuConsole I can screendump from."

This patch skips screendump for that kinds of vms to avoid the error
message and repeately sending the black images. On UI side, it changes
to display vm's icon instead of the black image if it doesn't have
video device.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>