bug fix: Do not store internal repository information
The former repository management implementation was storing the
repositories data internally. So it got the system data once and use it
on it. Which is wrong, as user can manage the system with other tools
and kimchi will show inconsistent data.

In addition to it, this patch also make many improvements:
  - Keep repository configuration file consistent on each actions
  - Use ConfigParser to add a new YUM repository
  - Let user specify a mirrorlist and a baseurl
    The mirrorlist can be used instead of or with the baseurl option.
    For reference: http://linux.die.net/man/5/yum.conf
  - Remove only the YUM repository section on config file instead of
    remove the while file
  - List all repositories (instead of only the enabled ones)

Signed-off-by: Aline Manera <alinefm@br.ibm.com>