VLAN: Do not allow bridge to be the trunk device
In Linux networking, a usual VLAN + bridge network setup is as the
following.

eth0 |-> eth0.10 -> br10 -> vnet101, vnet102, ...
     |-> eth0.20 -> br20 -> vnet201, vnet202, ...

While the eth0 trunk and VLAN interfaces provide the isolation service,
the bridges provide the switching inside the respective VLAN.

It's not very useful to have a bridge device as the trunk because the
VLAN interfaces on a trunk should be isolated.

This patch contains changes to the back-end and front-end.

The back-end checks if the front-end submits an invalid setup to use a
bridge as the VLAN trunk device and raises exception.

Before this patch, if the user creates a bridged network over an
existing bridge with VLAN tag set. The back-end just ignores the VLAN
and creates an ordinary bridged network. After this patch, the back-end
raises an exception with an explanation in this case.

The front-end also checks the currently selected interface type. If the
selected interface is a bridge, it disables the vlan check box,
otherwise it enables the vlan check box.

V2
Add front-end code to disable vlan check box if the selected interface
is a bridge.

V3
Clear the vlan check box and vlan id input box value when disabling them.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>