Remove debug report's file path from UI
Now the file path displayed on UI is "/data/debugreports/<reportname>/",
It's misleading because people could assume it's path on file system,
but it's the uri path in fact. We don't need expose the URI path because
the download button can help user to get the file. So this patch removes
it from UI.  And this patch also renames debug report's 'file' attribute
to 'uri', because we use it to represent the download uri path.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>