Template: Check if the host supports Spice
On some architecture of hosts, such as ppc64 hosts, QEMU is not compiled
with Spice. So we'd like to disable Spice on these hosts. Unfortunately
even QEMU is not compiled with Spice, 'qemu --help' still prints Spice
related help. In this patch we use ldd to determine if QEMU is linked to
libspice-server.so, and provide this information through the
/config/capabilities URI. The front-end code then checks the capabilities
and add or remove Spice from the graphic options respectively.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>