AsyncTask: Propagate cherrypy request information to worker threads
cherrypy.request is a proxy for cherrypy.serving.request, and
cherrypy.serving stores the request information on the thread local
storage. When new threads created by Kimchi code, the related request
information is not propagated to the new threads. The problem is that
exception message translation code needs the Accept-Language header or
session cookie to determine which language the user speaks. If the
request information is missing, it assumes the language is English.

Usually this is not a problem because most of the threads in Kimchi just
run in background, and the exception message is written to log, so
English in log is OK. However the AsyncTask catches the exception in
thread and reports to the front-end, and this is not friendly to the
user without proper translation.

This patch propagates the request information to the new threads started
by the AsyncTask, so exception message gets translated according to the
request header or session cookie.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>