Fix 'disk full' issue: Fix vms/screenshot db store/delete error handling
When creating, running or deleting a vm, it uses the database to store
vm and screenshot information. If the disk is full, Kimchi will raise
errors constantly and UI will be useless. This patch fixes this problem,
omitting from UI and logging in the backend. So, user is still able to
create, start and stop a vm.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>