Try to activate physical interface when create a network on it
Currently, the backend network interfaces API doesn't filter out
the down interfaces. So the user could create network on top of
a down interface. Kimchi should try to activate the interface if
it's down and report failure if the cable is not plugged.

In future, we could report the link status to front end and give
a prompt to user that it need takes several seconds to activate
the interface.