Use proper term "user name" instead of "user id"
In the Linux environment, a user ID is an integer code which identifies
a user; a user name is the human friendly text identifier of that user.
Kimchi uses both terms interchangeably.

Rename all occurrences of "userid" (and its variants) to "username" (and
its variants) in external UI messages and internal code.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@br.ibm.com>