change the stop to power off for VM in backend
"Stop" is obscure for user.

Currently, stopping guests in Kimchi is not graceful in the sense that
the guest becomes 'shut down' immediately. This could be be problematic
because guest file systems could become corrupted. Perhaps a warning
should be added before a guest is immediately stopped or reset.

So "power off" is more exact than "stop" in literal meaning.

The aciton power off a VM forcefully is same that we cut the physical
machine power. This action may produce undesirable results, for example
unflushed disk cache in the guest.

update API.md, backend source code and test case code.

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>