error when editing repository entry with no mirrorlist.
After edit a repository entry that contains only the baseurl field,
the new content of the entry will not be stored due to an error when
using '' string to the mirrorlist field.

 File "/home/pvital/git/kimchi-test-upstream/src/kimchi/repositories.py", line 267, in updateRepo
entry.mirrorlist = mirrorlist
File "/usr/lib/python2.7/site-packages/yum/config.py", line 102, in __set__
  value, str(e)))
ValueError: Error parsing "mirrorlist = u''": URL must be http, ftp, file or https not ""

This patch fixes the issue with a verification of the mirrorlist
before process it.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>