Fix kimchi vlan tagged bridge name
Before we use 'kimchi-{iface_name}-{vlan_id}' as the new created
bridge name.  It will exceed the maximum length of bridge name when
the interface name is too long. For example, when the udev
rule of renaming interface name by pci slot is enabled, the interface
name could look like 'enP3p9s0f2', which indicates the pci address of
nic device. According to systemd udev code, the prefix 'en' stands for
'ethernet'. We don't need preserve it in the bridge name. So this patch
change the bridge name convention to 'kb-{iface_name}[-8:]-
{vlan_id}'. It can keep the interface information with length less than
16 chars.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>