Github #329: Kimchid, config.py.in and server.py changes
Two new config options were added in config.py.in - proxy_port
and proxy_ssl_port - to allow more control in the ports that
the proxy will bind. The defaults proxy ports are 8000 and
8001 (ssl) and default kimchid port is 8010. The default proxy
ports were chosen to avoid editing the existing firewall
configuration (now that nginx is process being exposed).

Kimchid creates the proxy configuration according to user-defined
parameters.

server.py registers the terminate_proxy() method to a cherrypy
engine that fires when the server is shut down. The SSL logic
was moved to the proxy module at src/kimchi/proxy.py.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>