New domain state pmsuspended
libvirt introduces a new domain state pmsuspended to represent
the domain which has been suspended by guest power management,
e.g. (entered itno s3 state). Because a "running" state could
be confused in this case, one will see the guest is paused
actually while playing. And state "paused" is for the domain
which was paused by virDomainSuspend.

update this state in kimchi

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>