repository: Make checkbox clickable
A good usability feature of an HTML form is to click on an element's
label to activate the element itself. For example, when clicking on a
checkbox's label, the checkbox itself is checked, instead of requiring
the user to click on the small corresponding element.

Assign the checkbox "Repository is a mirror" to its corresponding label.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>