bug fix: Use secure cookies
Since this cookie does not contain the "secure" attribute, it might also
be sent to the site during an unencrypted session. Any information such
as cookies, session tokens or user credentials that are sent to the
server as clear text, may be stolen and used later for identity theft or
user impersonation.
Fix it.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>