Refactor code to trim file extension
In some parts of the code, it is necessary to remove the file extension
from a string. Currently, the last 5 characters are removed from
the text.

Remove the file extension by searching for the last '.' inside the
string. This approach works for other file extension lengths - even
though we will probably use only .html - and is more clear than having
a magic number ("5") inside the code.
Also, add a comment explaining what the substring method does in
that case.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>