bug fix: Redirect 3xx responses through proxy
When there is an redirection after POST/PUT/DELETE method, the proxy
must receive the response and redirect it properly to the correct server
and port.

The problem was identified when trying to rename a VM or template, or
even while cloning a template.

Fix it by adding new rules to nginx proxy configuration.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>