Revert "Enable encryption in vm VNC console connection"
This reverts commit 06238f292aaf236c456dcc50d0aea1efcbb05584.

This patch was exposing all the noVNC files and let websockify web server
render the noVNC page.
The websockify web server is limited and only renders content in a directory.

So if someone has the URL
https://host-ip:64667/vnc.html?port=64667&path=?token=my-vm&encrypt=1
he/she can access the VM console without Kimchi authentication.

The idea is let Kimchi render the noVNC page and also add authentication
to it.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>