Make use of the mini Web server in the websockify
Some browsers doesn't support well for the usage self-signed certs in the ssl
websocket connection. For details, please see:
https://github.com/kanaka/websockify/wiki/Encrypted-Connections

For chrome browser, the encrypted console connection should work after
you login with ssl connection. But for firefox, it does not show a
confirmation page for the user the accept the self-signed cert when
the HTTPs connection is started from a websocket.

So this patch makes use of the Web server in the websockify. The mini
Web server in the websockify can serve static contents like html, css
and js.
This patch add a simple HTMl file (console.html) to pages/websockify and have
websockify serve this file.

When the user clicks the console icon, Kimchi brings the user to
https://host:64667/console.html, which is served by websockify. Then
firefox would prompt a confirmation page for the self-signed cert. After
the user accept the cert, the user will be redirected to noVNC/SPICE page
provided by Kimchi server.

It is important to have Kimchi providing the noVNC/SPICE page to be able to
add authentication to console pages (vnc_auto.html and spice.html)

Signed-off-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>