Disable metadata log output in every guest refresh
The libvirt function "metadata" always displays an error message when
the VM does not have a metadata section in its XML descriptor. Kimchi
refreshes the VM information every few seconds and a lot of error output
is printed on the console. The real exception can still be logged by the
"try/except" block, so the error itself is not being silenced, only its
output.

Even though this patch has no relation to feature tests, the functions which
enable/disable the libvirt error messages are located inside the class
FeatureTests so we need to import them.

Silence the metadata error output when trying to read the metadata
information from a VM.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>