Remove button "Cancel" from modal dialog
Only one of the modal dialogs in Kimchi has a "Cancel" button along with
the form submit button. That button layout is inconsistent with the rest of
the application and it provides the same functionality as the "X" button
present at the top of every dialog.

Remove the button "Cancel" from the dialog "Edit template".

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>