Bug fix: Display the username on the header bar
The commit 85e3837 has introduced a bug where the username is not being
displayed on the header bar. Also, if the user hovers the mouse on the
drop down menu ("Help" - "About" - "Log out"), the element changes its
position by a few pixels.

Fix the bug mentioned above by setting the username to a browser cookie.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>