Add progress Event Handler for Asynchronized Tasks
Asynchronized tasks need continuously tracking until they are finished
to allow API consumer register a listener to take actions accordingly.
One of the common case is file uploading. When the user uploads a file
he expects there is a progress bar showing upload percentage.
  In our case, after we submit the generate debug report form, we will
close the generate window and add a "generating" line in the grid when
we receive the first response with HTTP status code 200. If the status
code is 4xx, we will not close the window because HTTP 4xx means there
is something wrong with user's input and we will keep the window there
to allow the user do some correction.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>