bug fix: redirect to the protected page after login
add a 'next' attribute in login html page.

if "next" is not None, means there is a html protected page user
want to access. after authentication success, redirect to it.

if "next" is None, then redirect to the last page recorde in th the
cookie.

Test this case:
1. input a protected page URL in the browser.
such as:
https://localhost:8001/spice.html?port=64667&listen=localhost&token=vmname&encrypt=1

2. input a wrong password or username:
it should report:
"The username or password you entered is incorrect. Please try again."

3. input the right password and username:
It should redirect to the former page.
https://localhost:8001/spice.html?port=64667&listen=localhost&token=vmname&encrypt=1

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>