Debug Report: Use Generated Time instead of Most Changed Time
Currently debug report generation time is ctime, which stands for change time.
Though getting creation time for file is dependent to platforms, and we'll use
mtime which stands for modification time and for our case because debug report
file will not be modified since it is generated. So mtime suits for our case.

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>