authorization: Get role according to tab instead of sudo rights
Kimchi will allow one user role per tab but it protects its resources
through API URIs.
That way we need to map API URIs per tab to get the user role.
Do that by adding a new parameter to UrlSubNode() to describe which tab
the URI is used.
And then use that information to get the user role for each request.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>