issue #374: Use base64 encoding to launch VM console
websockify.py uses urlparse.parse_qs() to parse token parameter in the
console URL

/vnc_auto.html?port=64667&path=?token=issue%2B374&encrypt=1

And even encoding the VM name using encodeURIComponent() it replaces the
plus sign for a white space,

{'token': ['issue 374']}

That way websockify will never be able to find the console port based on VM
name.

To fix it, use base64 for encoding the VM name while setting up the
console configuration and accessing it using noVNC or Spice.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>