authorization: Filter resources by users and groups
Currently, every user with 'admin' role can perform any operation on any virtual
machine.

In order to add more security, Kimchi will only allow users listed in
the VM metadata - along with those with 'admin' role - to be able to
perform actions on it. A VM may contain a list of system users and groups
associated with it. If a user is not listed to access a VM, they will
not be able to see it or to perform any operation on it.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>
Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>