List "admin" as a valid system user in mockmodel
The mockmodel environment reports a static list of users available on
the system. Currently, it does not include the fake user which is used
by the mockmodel authentication.

In order to make the code consistent, the user which authenticates to
the mockmodel environment should be listed as a valid system user.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>