authorization: Remove authorization config from UrlSubNode
UrlSubNode is used to automatically load the application configuration
and set kimchiauth tool when needed.

If we use it to also handle the authorization configuration, we won't be
able to specify different configuration to collection and its resource
as Kimchi uses the same base URL to both.

Example:
@UrlSubNode("vms", True, ["POST", "PUT", "DELETE"], 'guests')
It meant that all the methods listed were exclusive for admin users.
Which it is not correct, as a user assigned to a VM can also perform POST,
PUT and DELETE actions.

To be able to distinguish the configuration for resource and collection, the
autorization mechanism was moved to controller.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>