bug fix: Avoid equals sign in VM console URL
From python documentation, base64.urlsafe_b64encode(s) substitutes - instead
of + and _ instead of / in the standard Base64 alphabet, BUT the result can
still contain = which is not safe in a URL query component.
As base64 can work well without pad character (=), remove it.

The problem with equals sign was only identified on Spice connections.
noVNC can deal well with that.

For reference:
- https://docs.python.org/2/library/base64.html
- http://en.wikipedia.org/wiki/Base64#Implementations_and_history

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>