PowerPC bootable ISO detection code
PowerPC ISO images does not follow the same rules as Intel's.
Intel ISOs uses ISO9660 extension called 'El Torito', which
describes the format of the Boot Record. PowerPC firmware, on
the other hand, looks for a file '/ppc/bootinfo.txt' which
describes how to boot the image.

This patch creates a function that opens the ISO file without
mounting it and, following ISO9660 standards, read the filesystem
trying to find the boot file that enables PowerPC booting.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>