Increasing nginx proxy timeout
The 60 seconds nginx timeout is creating some issues with actions
that takes some time to complete, like creating a VM with a NFS
template. After 60 seconds without a backend answer, the user is
presented with a 504 Gateway Timeout error.

This patch increases the timeout to 600 seconds, 10 minutes. This is
enough time to process all the backend requests and ensure that a
504 error will only occur in a legitimate error in the backend.

The Timeout Monitor of cherrypy is also being disabled, since we're
using the nginx timeout and the session timeout instead.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>