UI enhancement: Request /config/capabilities as soon as possible
/config/capabilities was requested twice for the templates tab: one
request to identify if qemu has spice support and other one to identify
if the qemu has ISO streaming support.

Those requests were made right before using the information, causing a
delay on UI until get the server response.

For example, for distant servers, while creating a new template the
"Remote ISO" option was disabled at the first time and after some time it was
enabled.
The same behavior was identified while editing a template, as the graphics
options blinked to add the spice on it.

To fix it, request /config/capabilities when loading the templates tab to have
this information prior to edit or create a new template.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>