bug fix: Properly list host partitions for Ubuntu 14.04 server
While trying to create a logical pool in a Ubuntu 14.04 server I got the
following error:

[08/Aug/2014:15:52:38] HTTP Traceback (most recent call last):
  File "/home/alinefm/kimchi/src/kimchi/control/base.py", line 110, in lookup
    self.info = lookup(*self.model_args)
  File "/home/alinefm/kimchi/src/kimchi/model/host.py", line 274, in lookup
    return disks.get_partition_details(name)
  File "/home/alinefm/kimchi/src/kimchi/disks.py", line 180, in get_partition_details
    dev_path = _get_dev_node_path(majmin)
  File "/home/alinefm/kimchi/src/kimchi/disks.py", line 41, in _get_dev_node_path
    return _get_friendly_dm_path(maj_min)
  File "/home/alinefm/kimchi/src/kimchi/disks.py", line 33, in _get_friendly_dm_path
    with open(dm_name) as dm_f:
IOError: [Errno 2] No such file or directory: '/sys/dev/block/253:5/dm/name'

It is because, even the maj_min starts with '253:', the file does not
exist.
So check this file exists in order to use it, otherwise get the dev path
from uevent file.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>