Add persistent flag to VM info
Kimchi can manage guests not created by Kimchi. If a user creates a
non-persistent domain and uses the Power Off option, it will destroy
the user's domain. In order to warn users with non-persistent guests
on Power Off, this patch adds a 'persistent' field (like the one for
networks and storage pools) to a vm's JSON representation.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>