i18n support: Changed the file type of plugins/sample/po/LINGUAS
Changed the file type of plugins/sample/po/LINGUAS, that was a symbolic
link to po/LINGUAS, to be a "real" file containing only the three
languages (en_US, pt_BR, zh_CN) for the sample plugin.

This prevent the make command to use the Kimchi file to scan for other po
files in the sample plugin directory.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>