storagevolume: Add download progress to task
The download operation can take a long time, so it's nice to have some
way to track its current progress.

Update the task's message with the number of bytes downloaded. The
format is "<bytes downloaded>/<total bytes>". If the number of bytes
cannot be read from the remote URL, <total bytes> will be "-".

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>