bug fix: Properly set max body size to nginx proxy
To support uploading file to a storage pool, the max body size must be
larger enough to allow uploading a ISO or image file.
It was properly set on cherrypy server but as we use nginx as a reverse
proxy it also needs to have the same config. Otherwise, nginx will block
request with body size larger than 1M (the default value).

Also add test cases to cover those scenarios.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>