storagevolume: Use default value for param 'name' when appropriate
The parameter 'name' is only required when creating storage volumes with
the parameter 'capacity' (i.e. an empty volume). When creating volumes
with 'url' or 'file', the volume name should have a default value if 'name'
is not specified.

Use a default name when creating storage volumes from a file or from a URL.
The default name will be the file's or the URL's basename. If an unknown
volume creation method is used (i.e. if another method is added in the
future and this function is not updated accordingly), the default name will
be based on the current time.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>