mock: Do not write file when downloading a volume
The storage volumes created via download in mockmodel are being written
to the filesystem. As it is a mock environment, there's no real need to
write that file.
After running the current tests, the storage pool at "/var/lib/libvirt/images/"
has the volume "anyurl.wor.kz" left because of that behaviour.

Remove the file writing implementation when creating a volume via
download.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>