issue #415: Do not allow renaming a debug report to the same name
If the user tries to rename a debug report to the same name (i.e. not
actually rename...), Kimchi displays an error message indicating that
a debug report with that name already exists.

Check the new report name before renaming it, and if it hasn't changed,
display a message indicating that the name is the same as the
original one.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>