issue #462: Do not allow user to enter non-integer template disk size
If the user enters a non-integer template disk size (e.g. 10.5), guests
created with that template fail to start.

In order to simplify the user experience, all numeric values are
supposed to be consistent. Currently, all numeric values (except the
template disk size) are integer numbers, so it's more consistent for the
user to use integer numbers here as well.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>