issue #447: Check download URL prior to start Task
When user provides an invalid donwload URL, the error message
"Unexpected exception" was displayed without any more information.

On Kimchi logs, there is:

Error in async_task 1
Traceback (most recent call last):
  File "/home/alinefm/kimchi/src/kimchi/asynctask.py", line 72, in _run_helper
    self.fn(cb, opaque)
  File "/home/alinefm/kimchi/src/kimchi/model/storagevolumes.py", line 192, in _create_volume_with_url
    with contextlib.closing(urllib2.urlopen(url)) as response:
  File "/usr/lib64/python2.7/urllib2.py", line 127, in urlopen
    return _opener.open(url, data, timeout)
  File "/usr/lib64/python2.7/urllib2.py", line 404, in open
    response = self._open(req, data)
  File "/usr/lib64/python2.7/urllib2.py", line 422, in _open
    '_open', req)
  File "/usr/lib64/python2.7/urllib2.py", line 382, in _call_chain
    result = func(*args)
  File "/usr/lib64/python2.7/urllib2.py", line 1224, in https_open
    return self.do_open(httplib.HTTPSConnection, req)
  File "/usr/lib64/python2.7/urllib2.py", line 1186, in do_open
    raise URLError(err)
URLError: <urlopen error [Errno -2] Name or service not known>

It is because the download URL was not being checked prior to start the
Task and this case was not checked on Task handler.

To solver it, check if the download URL is accessible prior to start the
Task so user can change it accordingly.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>