Bugfix: Cancel option not working properly in New Storage Define
V3 -> V4:
Move the duplicate part to a function that reduce the duplication.

V2 -> V3:
Make the button style change properly with different storage pool types.

V1 -> V2:
Make the "Create" button change style and disable the input box only
when clicked on "ok" in "confirm" diaguage. The button and the input
style stay the same if user cancel the creating job.

This bug fix the defect that when adding a new logical storage pool, a
confrim message box show up and when canceling it, the inputbox as well
as button "Create" is still disabled.

Signed-off-by: Wen Wang <wenwang@linux.vnet.ibm.com>