pep8: Use blacklist instead of whitelist
Currently, the command "make check-local" runs the PEP8 verification
rules only on files listed in a configuration file. New files must
still be added manually and not all files are being checked.

List the files which *shouldn't* be checked against the PEP8 rules
instead of listing the files which should be checked. Also, fix the
errors on those new files.

Reasons to blacklist each file:

- src/kimchi/config.py and tests/test_config.py: those files are
  generated automatically by replacing strings inside them. We cannot
  guarantee that every line will still be shorter than 80 characters
  after replacing a text coming from an external resource.
- src/kimchi/websocket.py and src/kimchi/websockify.py: those files
  are from a 3rd party library and weren't developed by the Kimchi team,
  so we don't need to check their style.
- src/kimchi/i18n.py: this file only contains translatable strings. It
  would not be nice to wrap every line to 80 characters as they're not
  Python code but natural language text.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>