Disabling screen log in production environment
When building kimchi's RPM, some messages are getting
logged in /var/log/messages because cherrypy is logging
in the standard output. This happens because a function
in FeatureTests is enabling log.screen regardless of the
current environment. This patch makes a verification that
prevents the screen log to be enabled in the production
environment.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>