Update get_disk_xml() to get the device same according to bus and index values
When generating the disk XML, the most common information to generate the
device name are: 1) the bus - to identify the right prefix, and 2) the disk
index.
Add this logic to get_disk_xml() so when no 'dev' name is specified the
device name will be automatically generated according to bus and index
value.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>