VMsModel: Make _update_guests_stats() robust against race condition
"VMsModel._update_guests_stats()" is passed to cherrypy BackgroundTask
to run in a background thread. The main responsibility of this method is
to collect and update VM running statistics. The problem is that a
virtual machine can be destroyed and deleted when the method is running,
so in any part that deal with virtual machine in this method, it's
possible to trigger exception.

If you notice sometimes the following exception when runing test_model
tests, the problem described above is the cause.

Exception in thread Thread-4:
Traceback (most recent call last):
  File "/usr/lib64/python2.7/threading.py", line 811, in __bootstrap_inner
    self.run()
  File "/usr/lib/python2.7/site-packages/cherrypy/process/plugins.py", line 475, in run
    self.function(*self.args, **self.kwargs)
  File "/home/edward/src/kimchi/src/kimchi/model/vms.py", line 79, in _update_guests_stats
    dom = VMModel.get_vm(name, self.conn)
  File "/home/edward/src/kimchi/src/kimchi/model/vms.py", line 457, in get_vm
    raise NotFoundError("KCHVM0002E", {'name': name})
NotFoundError: KCHVM0002E: Virtual machine kimchi-ide-bus-vm does not exist

The cherrypy BackgroundTask will stop running once it detects exception
in the task method. So if we want to keep the "_update_guests_stats()"
running, we should swallow all kinds of exceptions in the method body.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>