ModelTests: Improve leftover cleaning
In ModelTests we use RollbackContext to register undo actions for
cleaning leftovers. The general principle for using RollbackContext is
that we should register an undo action right after the successful
allocation of a resource, otherwise we leave a window between the
resource allocation and cleanup action registration. If it happened to
raise an exception in this window, the allocated resource is not
cleaned.

In ModelTests, you can find that some test methods is not properly using
RollbackContext, they place assertions between resource allocation and
cleanup registration. This patch tries to adjust the order peroperly.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>