Make disk type an optional parameter on get_disk_xml()
There are cases in which the disk file does not exist but we know prior
to it what is the disk type so in this case the disk type should be read
from a parameter.

So only when no value for disk type is passed, we get the disk type according
to source path.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>