Set guest disk cache to none to support live migration
According to commit 707aee09 the configuration "cache=none" must be set
to allow live migration.
So do it on get_disk_xml() to make sure all guest disks will have this
configuration.

The test needed to be updated because libvirt uses O_DIRECT flag when
cache=none to read the disk file and it is not supported on tmpfs.

For reference: https://www.redhat.com/archives/libvir-list/2013-August/msg00783.html

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>