Fix content of the SW update list after disable a repository.
Kimchi lists all RPM packages to be updated when the Host tab is
selected, but even after disable some YUM repository and refresh the
content of this tab, the previous packages from the disabled repository
are still present in list.

This patch "deletes" the YumBase object after get the list of packages
to be updated. With this, every time the Host tab is selected, a new
list will be generated.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>
Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>