Add model function to wait for task
Sometimes we need to wait for a task to finish. Instead of writing the
waiting loop everytime, we can now use the function "wait".

Add the function "TaskModel.wait" which waits for a task to finish
for a maximum ammount of time (in seconds). If the task finishes before
the timeout expires, the function returns successfully; otherwise, an
exception is raised.
Also, update calls to an already existing function which does the same thing
in the test code, and replace its usages with this new function.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>