Clone storage volume
In order to clone VMs, we will need to be able to clone storage volumes.

Add a new method to clone a storage volume, e.g.:

POST /storagepools/<pool-name>/storagevolumes/<volume-name>/clone {'pool': 'another-pool', 'name': 'new-name'}

This method uses the parameters 'pool' and 'name' which specify the destination
storage pool and the volume name. If 'pool' is ommited, the new volume will be
cloned to the same pool as the original volume; if 'name' is ommited, a default
value will be used.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>