MockModel refactor: Create MockModel based on Model("test:///default")
The mockmodel was designed to provide a way to user tests Kimchi
without affecting the system (by "kimchid --test") and also to make the
tests easier to do.
But in fact, we have a bunch of code that completely differs from
the real model, so the developer needs to do 2 kinds of implementations
while developing a new feature.

This patch change MockModel to be Model("test:///default") and only
overrides what is not supported by the libvirt Test Driver.

It also fixes the test cases after that change: only delete a inactive
storage pool, the Test Driver already has on VM created on start up, the
ISO file must be an existing path, etc.

The run_tests.sh.in script was also updated to run the test_model.py
prior to the MockModel tests as the MockModel will override some Model
methods which can cause problems if the MockModel runs before Model.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>