Update clone test
The clone model test assumes that the default VM is running so it can
expect an exception when trying to clone that VM. But that's not always
the case. Also, the test doesn't wait for the clone operation to finish,
which sometimes may cause an error.

Make sure the VM used in the test is running before cloning it
(expecting an exception) and wait for the clone task to finish before
moving on to the next test.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>