Return empty dict when VM doesn't have current snasphot
The current API design is to return an error 404 when a current snapshot
doesn't exist on a virtual machine. But it was decided that the best
approach is to return an empty structure instead.

Return an empty dict instead of raising an error when a VM doesn't have
a current snapshot. Also, update the tests accordingly.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>