issue #438: Display nicer error message when starting network
If an error occurs when a network is being started (e.g. the host
firewall is disabled), a generic message is displayed.

Add a more specific error message for when a network fails to start. The
root reason is also displayed along the message.

Fix issue #438 ("Unexpected error when starting network interface with
firewall disabled").

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>