issue #526: Support updating name for VMs with snapshots
In order to update the name, the corresponding VM must be removed and
then added again. And a VM cannot be removed if it has snapshots. So, a
VM with snapshots fails to have its name changed.

Before removing a VM, a backup of the existing snapshots will be taken
so they can be recreated when the VM is added again.

Fix issue #526 ("Cannot edit VM with snapshots").