issue #504: Add details on how to install test dependencies
The package dependencies listed in docs/README.md are only required for
building, installing and running Kimchi. If the user wants to run the
tests, they need to install some extra packages which are not listed in
that file.

Update the file docs/README.md by adding the dependencies required for
testing.

Fix issue #504 ("check pyflakes installation").

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>