issue #553: Use required parameter in "snapshotLookupByName"
The function "snapshotLookupByName" has a required parameter in
older versions of libvirt. On newer versions, however, this parameter
is optional and it was causing an error on RHEL because it uses an old
version of libvirt.

Always use the parameter in function "snapshotLookupByName" with its default
value (0).

Fix issue #553 ("Get snapshots of vm raise 500 error").

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>