Reorganize the network tests
It creates 2 new files: test_network.py and test_mock_network.py with
the network test cases.

The only test case on test_mock_network.py is related to VLAN tagging
bridge as it is hard to test on Model environment because it needs an
active interface to create the VLAN tagging bridge.

MockModel was also changed to return the system interfaces as the
libvirt Test driver needs a valid one to create the VLAN tagging bridge
but it does not need to be active.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>