Change "_get_storagevolume" to static
The class StorageVolumeModel is instantiated during the creation of
StorageVolumesModel. Thus, the class StorageVolumesModel cannot have
an instance of StorageVolumeModel initialized in its constructor as well,
otherwise it will cause an infinite loop (i.e. StorageVolumeModel creates
an instance of StorageVolumesModel during its initialization, which creates
an instance of StorageVolumeModel...).

On future patches, we will need to use the function
"_get_storagevolume" (StorageVolumeModel) from inside
StorageVolumesModel, and due to the problem described above, it isn't
possible to have an instance of StorageVolumeModel inside
StorageVolumesModel.

Change the function "_get_storagevolume" to static, and rename it to
"get_storagevolume".

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>