Use 'bytes' as volume capacity and allocation unit
When creating a storage volume using the REST API
(i.e. "POST /storagepools/<pool>/storagevolumes {'capacity': 5}",
the parameters 'capacity' and 'allocation' must be specified in megabytes
(actually, mebibytes). This approach is not consistent with the REST
command "GET /storagepools/<pool>/storagevolumes/<vol>", which returns
the same values in bytes, and it also doesn't allow creating a volume
with a size not multiple of 1 MiB.

Update the REST command "POST /storagepools/<pool>/storagevolumes" so it
receives size parameters in bytes.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>