Run Model tests prior to MockModel tests
As described in commit fb498632, the Model tests must be run prior to the
MockModel tests as the MockModel will override some Model methods which can
cause problems if the MockModel runs before Model.

So rename the test_network.py to test_model_network.py and update the
run_tests.sh.in to run the Model tests first according to file name.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>