Bug fix: Allow clonning a volume from a pool with non-ASCII characters
All the data received from server is in unicode format, so we need to
encode to UTF-8 to get the right data value.

The functions StorageVolumeModel.get_storagevolume() and StoragePoolModel.get_storagepool()
used by clone() function also expects unicode values and then do the
encoding to UTF-8 so on clone() function there is nothing to do related
to encoding.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>