issue #512: Attach the function "showMessage" to the UI grid class
The JavaScript "class" kimchi.widget.Grid (which renders the packages
and repositories list) lacks the function "showMessage" (which
displays a message inside the grid). The function itself is already
declared but it isn't attached to the class, so it cannot be used. When
an error happens while reading the update packages or repositories list,
that function is supposed to be executed, however it doesn't exist in the
class and nothing is displayed.

Attach the function "showMessage" to its corresponding class so it can
be used correctly. In order to test this patch, create a failing
scenario to get the update packages list and open the tab "Hosts".
Without this patch, the update grid will be empty; with this patch, an
error message will be displayed in the update grid. On Fedora, the failing
scenario can be created by running "sudo yum update" (when there are
packages to be updated) and not confirming the update, so Yum leaves a
lock preventing other simultaneous operations to be executed.

Fix issue #512 ("Software updates" doesn't show available updates).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>