issue #545: Handle simultaneous authentication methods when updating VM permission
If a VM has one authentication method in its metadata (e.g. 'pam') and
a user sets some permissions with a different authentication method
(e.g. 'ldap'), then the group of the first authentication method is
lost and only the last type is kept. In other words, it is not possible
to have more than one authenticatin type in a VM's metadata.

Add the ability to handle different authentication methods
simultaneously when updating VM permission. Whatever is already set will
always be kept, even if the user chooses a different method.

Fix issue #545 (VM permission data is lost when changing authentication
method).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>