Assign HTML labels to their corresponding elements
Somea UI form fields have labels which aren't assigned to those fields,
which makes it more difficult to click and interact with them.

Assign the labels to their corresponding elements, so when the user
clicks on a label, the element also receives the click event. Also, some
indentation levels have been changed.

The elements which have their labels assigned to are:

- The text box where the user types the file path in "New Template" >
"Local image file";
- All form elements in "Templates" > "Actions" > "Edit", tab
"Processor";
- The checkbox used to select a device in "Storage pools" > "Actions" >
"Extend";

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>