Fix issue #620: Allow use iSCSI/SCSI volume on Template
When selecting an iSCSI/SCSI volume to create a Template, the disk size is
calculated according to the selected volume.

In that case only the volume information should be used:
{
  ...
  "disks":[
    {
      "volume":"unit:0:0:2",
      "index":0
    }
  ]
}

Also to get and display the disk size for user we should query the
selected volume.

This patch also created function to call GET /storagepools/<pool> API.
That way we can easily get the pool type without the need to query the information
on all pools returned by GET /storagepools API

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>