Fix test_model.py to reset Singleton classes when initializing the tests
Some FeatureTests functions depend on server to validate their result.
As CapabilitiesModel is a Singleton class it will get the first result from
FeatureTests which may be wrong when using the Model instance directly - the
case of this test_model.py
So clean Singleton instances to make sure to get the right result when running
the test_model.py tests.

This patch also created a setUpModule() and tearDownModule() to improve
performance while running the test_model.py tests.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>