Use more generic message in repo mirror list error
The current error message when the user provides an invalid mirror list
URL implies that the repository is DEB-based, but that error message is
used for other repository types as well (e.g. YUM).

Update the repository mirror list error message in order to be valid for
all repository types.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>