issue #461: Add 'metalink' support for YUM repositories
YUM repositories have a field called 'metalink', which is another
method to specify how to reach the repository data. Kimchi, however, is
unaware of that field.

Allow creating and updating a YUM repository with a metalink value, and
return the field 'config > metalink' when listing a YUM repository.

Fix issue #461 (repositories: Kimchi does not display metalink
information).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>