Move stats-related VM functions to VMModel
The variable "stats" and the functions "_update_guest_stats",
"_get_disk_io_rate", "_get_percentage_cpu_usage" and
"_get_network_io_rate", declared in src/kimchi/model/vms.py, are related
to a single instance of a VM, so it makes more sense to move them to the
class VMModel.

Move the variable "stats" and the stats-related functions from
global/static to VMModel.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>