Add libvirt-daemon-config-network package as Kimchi dependency
The libvirt-daemon-config-network package for Fedora/RHEL and openSUSE
installs /etc/libvirt/qemu/network/default.xml which creates the default
network.
So having this package as dependency eliminates the Kimchi needs to create on
start up.

This package is not required for Ubuntu, as it is already included in the
libvirt-bin package.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>