fix: Use correct path when setting 'ref_cnt' to a new volume
When a storage volume is created with the parameter 'capacity', the
parameter 'ref_cnt' is set on the storage pool path instead of the
storage volume path. 'ref_cnt's are only related to storage volumes, not
pools.

Use the storage volume path instead of the storage pool path when
setting 'ref_cnt' on a new volume.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>