Feature test to check support to memory devices
This patch adds a new feature test to check if the host libvirt
version supports memory devices attachment. This support is provided
since libvirt 1.2.14 and is the base to make memory hotplug work.
When libvirt does not support memory devices, it is going to fail when
we try to attach it, raising the libvirt.libvirtError:
"unsupported configuration: unknown device type 'memory'"

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>