Add maxMemory and numa configuration to guest xml
In order to support memory hotplug, guest must have maxMemory and NUMA
configured in the xml.
For maxMemory, this patch changes template with this information and
the xml generation at guest creation time, adding maxMemory equals to
host total memory and memory slots as the integer number of GiB that
fit inside maxMemory (by design users will be allowed to add only
memory chunks of 1GB).
For NUMA, this patch adds the simplest configuration possible, creating
only one node with all vcpus and memory set in the template.
VM update function was changed in order to properly update memory, numa
and cpu informations. This patch also provides a mechanism to set these
parameters in old vm xml, user just have to update the memory offline
once, then the memory hotplug is going to be set and enabled.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>