Getting default disk format from template.conf
If the user change disk format in template.conf, templates are going to
be created correctly and vm xml as well. But, when the disk is created,
the format is always qcow2.

This patch considers the settings in the template.conf to define
the default volume format when creating the disks. User defined settings
will override any default (unless the storage type is logical,
iscsi or scsi. In those cases the format will always be 'raw').

If the user does not provide any format and template.conf does not
declare any default as well, the default value to be used as volume
type will be 'qcow2' (considering the storage type restriction
mentioned above).

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>