Improve code to get default disk format for VMTemplate
The default template configuration is defined in the template.conf file.
This file is parsed on server start up and the result is always returned
in the osinfo.lookup() method. That means when creating a new
VMTemplate instance it will already contain all the default values to
use for the Template creation.
The disks are in a dict format like below:

[{u'index': 0, u'size': 10, u'format': u'qcow2'}]

Each disk has its own size and format and we just need to use them.
We don't need to parse the template.conf file again in order to get the right
values.

Also adjust the tests according to those changes.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>