Fix bug #450: Allow creating guest with non-ASCII characters in name
As discussed in the github (https://github.com/kimchi-project/kimchi/issues/450),
the idea to solve this problem is always save the guest name with ASCII characters
and use the <title> tag to handle the non-ASCII name.

As the <title> tag is wide used for libvirt guests and Kimchi already has its own
<metadata> section in the guest XML, I decided to include the non-ASCII name
in the <metadata> as well, to avoid any conflicts with the <title> tag.

When the guest name has non-ASCII characters, it is encoded to base64 and the
unicode name is stored in the <metadata> as <name>.

The backend is responsible to hanble those values, which means the UI
keeps working with the non-ASCII name without problems.

Another known issue is related on how Javascript works with base64 and non-ASCII
characters which is preventing Kimchi to properly display the guest console.
The problem is because atob() and btoa() do not work well with non-ASCII
characters which turns into a wrong base64 result.
For example, "café" would be "Y2Fmw6k" but btod() turns it into "Y2Fm6Q"
This is another issue and related to UI. Another patch to fix it will be
send soon.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>