PCI filter not in sync with device listing
When editing the VM, in the PCI tab, removing a device or adding a pci
to the list, only the '+' or '-' icon changes but the list doesn't get
refreshed.  This behavior is ok when in the 'All' filter.  However, when
in the 'ToAdd' filter, the device should no longer be shown after it has
been added and same behavior should occur when in 'Added' filter after
it has been removed.

Signed-off-by: Socorro Stoppler <socorrob@us.ibm.com>