New UI updates
- Adding font-awesome and open-sans CSS files to UI

Adding the CSS files pointing to the correct font path in the server.
This patch also removes Fontello references from the project and updates the
config and test_config files with the correct path for these font families CSS.

- Moving JS libraries to new folders and added license files

This commit updates jquery and jquery-ui to the latest compatible versions
(v1.11.3 and v1.11.4) and adds all the JS libraries that are required in order
to run and develop the new-ui.
Updated the old-ui login.html.tmpl and wok-ui.html.tmpl to point to correct Js
path otherwise the CherryPy would throw a 404 error when loading the pages.
Moved all libs/themes files to libs/jquery-ui/themes/ and added a copy of
jquery-ui license to this folder.
All the new libraries where added to ui/libs folder and configure.ac was
properly updated to read the new Makefiles.

*New-ui related:
   *Removed logo.ico and updated with favicon.ico in high resolution
    (name convention for IE9 compatibility). This change can be spotted in
    the browser tab bar with the old-ui.
   *Updated opensans.css to point to local font family when the user already
    has it installed.

- Adding new webapp icons for mobile devices and Windows

This commit adds Kimchi logo in HD as a bookmark icon for modern browsers,
mobile devices  menus and Windows 8/8.1/10 start menu tiles.

This update also includes two SVG icons for the new-ui with embedded licenses.
*spin5.svg is part of Fontellico font and it is distributed under CC BY 3.0.
*edit-alt.svg is part of Elusive icons and it is licensed under SIL OFL and CC BY 3.0.

Signed-off-by: samhenri <samuel.guimaraes@eldorado.org.br>