Use locks to prevent concurrent updates to VMs
If several users try to update the same VM at the same time, an error
may happen due to concurrent access.

Define a dictionary containing one lock per each VM, and use it when
updating the VM: the lock is acquired before the update operation begins
and it is released when the operation finishes (whether by success or
failure).

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.vnet.ibm.com>