imageinfo.py: Don't fail template creation if unable to detect OS
This patch slightly alters the behaviour when creating a template
from an exisiting disk image (i.e. qcow2).  If guestfs is unable
to detect what distro, version, etc is within the image, it should
not be a catastrophic error.  In fact, the vmtemplate.py handles
the return of distro == 'unknown' already by creating a
uuid for it.

This is important for operating systems like Atomic (RHEL, CentOS,
and Fedora) because they have a non-standard partition and disk
layout that guestfs cannot readily figure out.