Live migration backend: unit tests
These unit tests requires an additional variable passed in the
command line, otherwise it won't be possible to test the
migration types. Example:

$ sudo KIMCHI_LIVE_MIGRATION_TEST=1.1.1.1 ./run_tests.sh test_livemigration

It also depends on the common pre-requisites of migration:

- valid remote host
- password-less root login at the remote host

The unit test will silently skip if the KIMCHI_LIVE_MIGRATION_TEST is not
defined to avoid tampering with the common usage of ./run_tests.sh script.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>