Issue #751: Fix guest memory utilization calculation.
The return of Pyhton's function memoryStats() returns the guest's memory
utilization statistcs by the Libvirt API virDomainMemoryStats. However,
the content of the return (a dictionary) depends on the combination of
host and guest OS.

This patch fix the issue of not returning the curent utilization of
guest memory when host and guest are running different OS and not all
statistics are returned.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>