Upgrade Kimchi objectstore content.
With the new structure of Wok and Kimchi, the content of the objectstore file
needs to be updated to reflect the new paths of icon and storagepool present
in template's and VM's information.

This patch adds a step to execute an update of these contents in the previous
objectstore file (older than Kimchi 1.5.1) during the startup of Kimchi 2.0.0
(or higher) plugin.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>