Fix issue #734: Update test case as libvirt issue was fixed
Due a libvirt bug it was not possible to decrease the storage volume capacity,
so the test case was updated accordingly.
As libvirt bug is fixed now, remove the hack in the test case.
For distros using older libvirt version, the test case may fail but the
error can be safely ignored.

For reference: https://bugzilla.redhat.com/show_bug.cgi?id=1021802

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>