WoK: control/base.py: _generate_action_handler_base changes
The function _generate_action_handler_base was enforcing
all parameters passed in the 'action_args' array as if
they were obligatory, disregarding any setting of the
API.json of the plug-in.

This patch makes this process more lenient, adding
only the parameters that exists in the request. The API.json
of each plug-in will handle the parameters usage.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>