WoK: making /tests/utils.py available to all WoK plug-ins
There are several common functions such as run_tests,
get_free_port and running_as_root that should be available
to all plug-ins. This patch enables the module /tests/utils.py
of WoK to be used by any unit test of any WoK plug-in. There's
also a small tweak in utils.py that allows the retrieval of
its 'fake_user' attribute.

This patch is part of the solution of Kimchi Github #760.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>