Issue #737: Wok base framework calls resource model's lookup twice
In case of GET lookup was called twice.
Once before is_authorized() and then in self.get().
This added overhead to system when lookup() is called
for each value in list from get_list() of Collection.
So to avoid this, removed lookup() call from get().
And updated code to call self.lookup() before self.get()
in resource's update() and collection's create().

Signed-off-by: Archana Singh <archus@linux.vnet.ibm.com>