Support Linux Bridge creation
This commit adds a new option to bridged network creation
API: "macvtap". When it is chosen, the usual behavior for
option "bridge" will be applied.

This commit also changes the behavior for the already
existing "bridge" option. Now, when that option is chosen,
a Linux bridge will be created, not a macvtap one.

Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>
Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>i