Update VCPU by using libvirt function
Currently, the VCPU count is updated by editing the configuration XML
(tag: <vcpu>). However, in some cases, editing that tag will only update
the maximum VCPU count, not the current one. For example, if the VM has
the following XML tag:

  <vcpu current='2'>8</vcpu>

and the user updates the VCPU value to 10, Kimchi will update the
XML tag to:

  <vcpu current='2'>10</vcpu>

Use the libvirt function "setVcpusFlags" to update the current and
the maximum VCPU values to the one specified by the user.

  * This patch also changes vcpu setting when guest is being
  created. Now 'current' is always set and is the total of cpus
  from the Template and maxVcpu is the >= 255 or (sockets * cores
  * threads) in PPC. In x86 it is the value of libvirt getMaxVcpu.

  * Change NUMA management. Sets all cpus to 0. There is not
  impact to the guest.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>
Signed-off-by: Crístian Deives <cristiandeives@gmail.com>
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.vnet.ibm.com>
Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>