Remove 'stogarepool' referencies from Templates code backend
It is not necessary to pass the parameter 'storagepool' when create a
new template or update one. User is going to set a pool/name for each
disk instead.
This patch remove the legacy code related to 'storagepool' in templates.
It also fixes minor issues found and improve some checkings.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>