Fix test_async_tasks testcase.
When executed all Kimchi tests (by run_tests.sh), the test_async_tasks is
falling due to wrong check of the number of the current task added to be
executed. This happens because the control of the task_id is handled by a
global variable in Wok, increased by 1 each time a new task is added.

This patch uses the wok.utils method get_task_id() to get the current value
of the task_id controlled by Wok.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>