Remove required authentication from /
Each plugin is able to specify if an URI needs or not authentication and
it must be per URI.
Setting / to require authentication to the whole application will lead on UI
issues and the frontend will not be able to load HTML content as it is
protected behind authentication.

This problem was identified when accessing an specific URI (for example,
https://localhost:8001/#plugins/kimchi/tabs/guests) without
login into Wok. The expected behavior was the user to be redirected to
the login page, but as the / was protected, a blank screen was shown.
So fix it, by removing authentication from / and keep it for specific
URIs.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>