Update Kimchi objectstore versioning.
With the new structure of the objectstore changed in Wok 2.0.0, the schema of
the objectstore file needs to be checked and, if necessary, updated to reflect
the new structure.

This patch removes from Kimchi the methods used to check the objectstore schema
and upgrade it if necessary. These methods are now present in wok.plugins and
are common to any Wok plugin.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>