Issue #753: Adds max_memory checkings and settings properly
When Kimchi creates a new template, it is recording the value of
max_memory. This value is related to host and guest memory assigned or
updated. If the value remains in objectstore, Kimchi will always use
it, what is wrong, since host total memory or guest memory can change,
and this should update max_memory value as well.

This patch includes a new memory test during Template creation.
An error will be raised if memory greater than max memory allowed.
It also adds max memory settings to vmtemplate.py

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>