Fix add storage device to VM using existing disk
This patch fixes the error of not being able to attach a
storage device to a VM using an existing disk.  The problem
was due to a change in the html file, the name attribute
of the select component no longer exists which was still
being referenced in the code.

Signed-off-by: Socorro Stoppler <socorro@linux.vnet.ibm.com>