Changing network API 'interface' to array: docs and API
Changed docs/API.md, i18n.py and API.json to reflect the new API.

interfaces[0] (the first element of this array) will have the
same effect as the former 'interface' parameter.

This change is necessary to better adapt the backend to networks
that can be created with more than one interface.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>