VEPA network support - models and xmlutils changes
-model/networks.py: added 'vepa' support by creating a new
method to verify the availability of all interfaces and
preset the parameters for the xml creation. Edited the lookup()
method to add the interfaces retrieved by the VEPA xml declaration,
keeping the same behavior for the other networks.

- xmlutils/network.py: edited the _get_forward_elem() method
to add all the interfaces that belongs to a 'vepa' type network.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>