VEPA network support: UI changes
Pinpoint UI changes to add a new network type named 'vepa'
with the same capabilites as the macvtap network.

At this moment, the UI does not support multiple physical
devices to be declared in the VEPA network. This will require
additional UI changes in the future (no backend changes
needed).

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>