Bug fix: Remove storage volume file while removing the storage volume
When removing a storage volume, it was only removed from a libvirt
perspective, ie, the storage volume file kept in the system.
It may confuse user as he/she will not be able to create a new storage
volume with the same name from the storage volume removed before as the file
exists in the system.
To avoid it, remove the file from the system after removing it from a
libvirt perspective.

This issue was identified due a failure in the test suite.
2 files related to storage volume tests were kept on the system causing
errors when running the test suite multiple times.

To verify this patch, please, make sure to do not have any leftovers in the
'default' storage pool directory (/var/lib/libvirt/images).
After that, run the test suite mutliple times and you will see no
leftovers will be in the system when the test completes.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>