Add function 'get_next_clone_name'
This patch moves the function 'get_next_clone_name' from Wok to Kimchi.

The function adds, originally, an incremental number in each clone name,
however this approach causes a race condiction if Kimchi tries to clone
guest multiple times, sending multiple request at once.
If clone requests are done so fast, it is possible that the plugin does
not send all resource names in all_names parameter, then the function is
going to select the same name for 2 or more clones, which is going to
cause errors.

To avoid this problem, it is better to use timestamps in the clones
names instead of incremental numbers. Then it guarantees names will
be unique. Supporting timestamps in microsecond precision, avoids this
type of error.
New clone name will be "basename-clone-timestamp-name-suffix".

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>