Issue #788: Network information mismatch in template creation
This commit fixes Issue #788.
- If a Template is using a network, Kimchi will prevent this network from being
stopped. However the user can still stop it using virt-manager.
- Now if a network is disabled and the user edits a Template, the user can see
the current network associated with this template: the dropdown will show the
current network + all the active networks in the list.
- If only 'default' network is created and it is stopped by virt-manager, when
the user attempts to edit the Template interfaces it still shows the default network.

This commit also fix a small glith in the UI when trying to enable a network.

Signed-off-by: Samuel Guimarães <sguimaraes943@gmail.com>