Issue #838: Extend memory hotplug feature test
If QEMU version is < 2.1, qemu is not going to start guest which have
maxmemory tag configured in the XML. Kimchi adds this tag by default in
order to allow memory hotplug.
In other words, in order to allow memory hotplug, versions must be
libvirt >= 1.2.14 and qemu >= 2.1. Libvirt was already being tested by
the feature test, but qemu did not. Distros as Red Hat 7 and CentOS
still are in qemu 1.5.
This patch extend the feature test and do not create maxmemory tag in
order to avoid problems in those distros.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>