Fix memory value return when hotplug memory devs
There is an issue when Kimchi returns the value of memory, if the guest
is running. It is returning the current memory value, instead of the
total memory, which includes the value of the memory devices added.
This patch fix this problem returning the total value: current memory
+ total of memory devs.
This patch computes the total of memory considering different memory
device sizes and units.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>