Avoid race condition in vm lookup
In guest lookup, libvirt domain is instanciated and then a info()
function is called, but it is possible that the guest be removed between
these two operations, raising an exception that is not treated.
This patch fix this issue, catching the exception, logging and returning
the correct error to the request.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>