Fix memory return in vm with memory devices hotplugged
Libvirt is updating 'currentmemory' and 'memory' if its values are
equal, when user attaches a memory device. So, we can use the function
maxMemory() of Libvirt API to retrieve the value of memory.
Kimchi does not use currentMemory tag, but once it is the amount seen
inside a running guest, if it was changed outside kimchi, we must
sum the memory devices values (if any) to have the right value seen
inside guest.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>