Make Cheerypy up if not able to connect to libvirt
Modified libvirtconnection.py to not stop Cherrypy server when not able to
connect to libvirt. Now, a notification message is displayed in UI to warn user
about libvirtd missing connection, making the web server available to provide
the other plugins from wok.

This patch is part of the solution for Issue #201

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>