Add support to Libvirt Events.
This patch adds support to handle in Kimchi any Libvirt Event, just by adding
a callback to process the event and register the event with the callback.

None event is being registered by this patch and they need be added by developer
by demand in the code. To see how register an event, the unit test
test_libvirtevents.py register two events and then use them in the code.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>