Load Kimchi when started by command line and libvirt is not running
Kimchi is able to detect and continue running if libvirt stops and back.
However if libvirt is not up and host admin tries to start Wok, Kimchi
is going to crash in feature tests and when loads models.
This patch fixes this issues, then Wok does not break and is able to
load and start Kimchi. When libivirt service is back, Kimchi is able to
run feature tests and set the capabilities again, restoring all
features.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>