Clean FEATURETEST_VM effectively.
FEATURETEST_VM is defined at Kimchi start up to check if few QEMU and Libvirt
features are enabled in the host system. Some bugs reported the existence of
this VM running in the system, even after the Kimchi Feature Tests have
completed.

This patch adds a cleaning step before start the Kimchi Feature Tests. In
addition, modified the has_mem_hotplug_support() to not use RollbackContext and
guarantee that FEATURETEST_VM will be undefined.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>