Fixed Storage Volume upload error message when switching tabs
This commit fixes a bug that would occur if the user starts uploading a volume
and switches to a different tab or select a different Storage Pool and starts a
new upload while the previous task is still in progress. It would return an error
message as "KCHPOOL0002E: Storage pool undefined does not exist" and stop the
upload progress but the upload task would keep returning "running" until Wok
server is finished.
This was happening because the global variabel kimchi.selectedSP was sent as
an argument to the functions kimchi.createVolumeWithCapacity() and
kimchi.uploadVolumeToSP().

The solution was creating a copy of the current kimchi.selectedSP value in
each kimchi.sp_add_volume_main() function call.

Signed-off-by: Samuel Guimarães <sguimaraes943@gmail.com>