Save view for Guests tab
This patch addresses the issue of the last view of the Guests tab not
being remembered either after switching tabs or exiting Kimchi.  When
switching view from List to Gallery (or vice versa), the next time the
user goes back to the Guests tab, whatever view it was last on should
be what is shown when going back to the Guests tab.

Signed-off-by: Socorro Stoppler <socorro@linux.vnet.ibm.com>