VM migrate: generic remote path check
The method _check_if_path_exists_in_remote_host was using
'test -f' to check if a file exist in the remote host.
Hovewer, when using iSCSI disks, this path is a symlink
instead of a regular file. This results in failure, making
the code wrongly go to the 'non-shared' migration scenario.

This patch changes the 'test' flag to '-e', which will simply
check for the existence of the path in the remote host. There
is no need to distinguish between whether the path is a file
or a symlink in this method - we just want to know if the path
is valid in the remote host.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>