Save last view for templates - fix for issue# 799
This patch addresses github issue# 799.  The last view of the Templates tab
is not being remembered either after switching tabs or exiting Kimchi.  When
switching view from List to Gallery (or vice versa), the next time the
user goes back to the Templates tab, whatever view it was last on should
be what is shown when going back to the Templates tab.  Thanks to samhenri
in helping resolve the last issue.

Signed-off-by: Socorro Stoppler <socorro@linux.vnet.ibm.com>