Issue #317 Inconsistent button status when adding or creating new resources
- Addressed last comment by pvital in github issue in that only
the Network tab now has this problem in that the 'Create' button
is enabled after clicking on   Add a Newtwork. It should be
disabled by default until the minimum of the required fields are
filled. In the Network case, network name field (as long as valid)
will enable the Create button as soon as at least one character
is typed in.

Signed-off-by: Socorro <socorro@linux.vnet.ibm.com>