Issue #982 - Fix broken testcases.
Several testcases started to fail in Fedora 24 after commit 24cc635 added
a solution to create a temporary file to represent the Virt Viewer script
contents in mockmodel.py. The solution is not able to create the directory
in where this temporary file will be created and is failing.

This patch adds a control to make sure the directory exists.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>