Adding libvirt remote connection verification
Depending on the remote libvirt configuration, setting a password-less
login between the hosts isn't enough to ensure a password-less API
connection to the remote libvirt service.

This patch adds a virsh verification to ensure that this password-less
connection can be estabished before moving further with the live
migration process. If the verification fails, a proper error message
is given to warn the user about the need of making a manual configuration
in the libvirt service of the remote host.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>