Adding 'sockets' field in the topology of Templates
Following the changes of bug fix #1072, this patch adds the
'sockets' values to be editable in the CPU topology. Since
sockets is now a component of the calculation of Max CPUs of
a VM, it makes sense to allow user control of it.

As with cores and threads, the defautl value of sockets is
returned by the cpu_info API.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>