Changing 'threads' to be a free number field
This patch removes the limitation, in both backend and frontend,
of the value 'threads' in the topology to be limited by
the threads_per_core value of the host.

This limitation was put in place to help the user pick a
'threads' value that provides a good performance, but it ended
up being a problem to support VMs what were created outside of
Kimchi and didn't impose such limitations.

To compensate, a help text was added to warn the user about the recommended
'threads' values.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>