CPU Hot plug/unplug: ui changes
This patch makes the following changes to enable CPU
hotplug/unplug on a running/paused guest:

- if the guest is running, lock all fields but the current CPU
in the Edit Guest dialog. A message is displayed informing
the user that on a running or paused guest the only CPU
setting that can be changed is the current CPU value;

- in the form validation processorSubmit function, discard
every info but the current CPU value if the guest is running
or paused.

Changes in guests.dita were also made to add information about
the guest support for this feature.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>