Fix issue #1069: Allow user specifies the Template name when creating it
When user creates a Template, now he/she can specify the Template name
otherwise, it will be automatically chosen.

The Template name is not taken in count when the user selected more than
one Image Template. In that case, a name will be automatically chosen
for each Image Template selected.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>