Bug fix: Set default host value while generating the virt-viewer config file
The test case was failing as below:

======================================================================
ERROR: test_vm_virtviewerfile_vmlifecycle (test_model.ModelTests)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/mock/mock.py", line 1305, in patched
    return func(*args, **keywargs)
  File "test_model.py", line 571, in test_vm_virtviewerfile_vmlifecycle
    vvmodel.lookup(vm_name.decode('utf-8'))
  File "/home/alinefm/wok/src/wok/plugins/kimchi/model/virtviewerfile.py", line 133, in lookup
    file_path = create_virt_viewer_file(name, graphics_info)
  File "/home/alinefm/wok/src/wok/plugins/kimchi/model/virtviewerfile.py", line 82, in create_virt_viewer_file
    {'name': vm_name, 'err': e.message})
OperationFailed: KCHVM0084E: KCHVM0084E

Because the virt-viewer config file generation was relying on Wok
configuration to get the host information but the host information is
not part of Wok anymore. Instead of that, it is being configured on
nginx. Set it to 'localhost' when that information is not part of
request headers.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>