Move Kimchi specific functions from gingerbase.disks to Kimchi
The idea behind this patch is to eliminate the Ginger Base dependency.
Some functions in gingerbase.disks are for only Kimchi matters.
Others are shared with Ginger and can not be rewritten using PyParted,
each means they will be place on Kimchi and Ginger Base.
As this code is mature enough, further changes will have few impact and
the gain to incorporate Ginger Base into Ginger and eliminate a Kimchi
dependency is bigger.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>