Change VNC & Spice console URLs to reference web interface port
This change allows the user to place a reverse proxy in front of
kimchi's built-in nginx proxy. The user-added reverse proxy can then
handle translation to the port exposed by kimchi/nginx.
Specifically, this patch uses the javascript location object to get
the current port and protocol used to connect to kimchi/wok.
Without this change, kimchi/wok serves up the wrong url (and port)
for the noVNC and Spice consoles when behind a reverse proxy.

Signed-off-by: Galen Pospisil <galen.m.pospisil@gmail.com>