Fix issue #1157: Consider DOS/MBR mime type as ISO file
While creating a Template, the user can provide an ISO file or an Image
file. Kimchi relies on python magic to get the file mime type and it was
only considering ISO 9660 CD-ROM as valid for ISO images and it was fall
backing to an Image file and failing .
Add DOS/MBR as well to fix that problem.

Signed-off-by: Aline Manera <aline.manera@gmail.com>